requirements.txt
Pillow
matplotlib
opencv-python
numpy
tqdm
torch
torchvision